Update AB main experiment with previously made changes
I don't know what `main-changed.ts` did but I assume it was just some
slight changes to the Attentional Blink experiment.
